+++ packages/manager/src/features/Kubernetes/KubernetesClusterDetail/NodePoolsDisplay/NodePoolsDisplay.tsx
@@ -5,13 +5,14 @@ import {
 import classnames from 'classnames';
 import * as React from 'react';
 import { useSelector } from 'react-redux';
+import { Waypoint } from 'react-waypoint';
 import AddNewLink from 'src/components/AddNewLink';
 import Paper from 'src/components/core/Paper';
 import { makeStyles, Theme } from 'src/components/core/styles';
 import Typography from 'src/components/core/Typography';
 import ErrorState from 'src/components/ErrorState';
 import Grid from 'src/components/Grid';
-import { ExtendedType } from 'src/features/linodes/LinodesCreate/SelectPlanPanel';
+import { ExtendedType } from 'src/store/linodeType/linodeType.reducer';
 import { useDialog } from 'src/hooks/useDialog';
 import useFlags from 'src/hooks/useFlags';
 import useLinodes from 'src/hooks/useLinodes';
@@ -99,6 +100,13 @@ export const NodePoolsDisplay: React.FC<Props> = props => {
   const deletePoolDialog = useDialog<number>(deletePool);
   const recycleAllNodesDialog = useDialog<number>(recycleAllNodes);
   const recycleNodeDialog = useDialog<number>(deleteLinode);
+
+  const [numPoolsToDisplay, setNumPoolsToDisplay] = React.useState(25);
+  const handleShowMore = () => {
+    if (numPoolsToDisplay < pools.length) {
+      setNumPoolsToDisplay(Math.min(numPoolsToDisplay + 25, pools.length));
+    }
+  };
 
   const [addDrawerOpen, setAddDrawerOpen] = React.useState<boolean>(false);
   const [resizeDrawerOpen, setResizeDrawerOpen] = React.useState<boolean>(
@@ -193,6 +201,8 @@ export const NodePoolsDisplay: React.FC<Props> = props => {
     });
   };
 
+  const _pools = pools.slice(0, numPoolsToDisplay);
+
   /**
    * If the API returns an error when fetching node pools,
    * we want to display this error to the user from the
@@ -254,7 +264,7 @@ export const NodePoolsDisplay: React.FC<Props> = props => {
         ) : (
           <Grid container direction="column">
             <Grid item xs={12} className={classes.displayTable}>
-              {pools.map(thisPool => {
+              {_pools.map(thisPool => {
                 const { id, nodes } = thisPool;
 
                 const thisPoolType = types.find(
@@ -279,7 +289,13 @@ export const NodePoolsDisplay: React.FC<Props> = props => {
                   </div>
                 );
               })}
+              {pools.length > numPoolsToDisplay && (
+                <Waypoint onEnter={handleShowMore} scrollableAncestor="window">
+                  <div style={{ minHeight: 50 }} />
+                </Waypoint>
+              )}
             </Grid>
+
             <AddNodePoolDrawer
               clusterLabel={clusterLabel}
               open={addDrawerOpen}