+++ packages/manager/src/features/NodeBalancers/NodeBalancerDetail/NodeBalancerConfigurations.tsx
@@ -976,6 +976,7 @@ class NodeBalancerConfigurations extends React.Component<CombinedProps, State> {
       checkPathLens: lensTo(['check_path']),
       portLens: lensTo(['port']),
       protocolLens: lensTo(['protocol']),
+      proxyProtocolLens: lensTo(['proxy_protocol']),
       healthCheckTypeLens: lensTo(['check']),
       healthCheckAttemptsLens: lensTo(['check_attempts']),
       healthCheckIntervalLens: lensTo(['check_interval']),
@@ -1030,11 +1031,13 @@ class NodeBalancerConfigurations extends React.Component<CombinedProps, State> {
           port={view(L.portLens, this.state)}
           onPortChange={this.updateState(L.portLens)}
           protocol={view(L.protocolLens, this.state)}
+          proxyProtocol={view(L.proxyProtocolLens, this.state)}
           onProtocolChange={this.updateState(
             L.protocolLens,
             L,
             this.afterProtocolUpdate
           )}
+          onProxyProtocolChange={this.updateState(L.proxyProtocolLens)}
           healthCheckType={view(L.healthCheckTypeLens, this.state)}
           onHealthCheckTypeChange={this.updateState(
             L.healthCheckTypeLens,