+++ packages/manager/src/features/ObjectStorage/BucketDetail/BucketAccess.tsx
@@ -1,0 +1,7 @@
+import * as React from 'react';
+
+export const BucketAccess: React.FC<{}> = _ => {
+  return <h2>Bucket Access</h2>;
+};
+
+export default React.memo(BucketAccess);