+++ packages/manager/src/features/linodes/LinodesCreate/LinodeCreate.tsx
@@ -64,6 +64,8 @@ import {
 } from './types';
 import { Tag } from '@linode/api-v4/lib/tags/types';
 import Notice from 'src/components/Notice';
+import SMTPRestrictionText from 'src/features/linodes/SMTPRestrictionText';
+import { filterCurrentTypes } from 'src/utilities/filterCurrentLinodeTypes';
 
 type ClassNames = 'root' | 'form' | 'stackScriptWrapper' | 'imageSelect';
 
@@ -132,6 +134,7 @@ interface Props {
   resetCreationState: () => void;
   setBackupID: (id: number) => void;
   showGeneralError?: boolean;
+  setVlanID: (id: number | null) => void;
 }
 
 const errorMap = [
@@ -142,7 +145,8 @@ const errorMap = [
   'region',
   'root_pass',
   'stackscript_id',
-  'type'
+  'type',
+  'interfaces'
 ];
 
 type InnerProps = WithTypesRegionsAndImages &
@@ -265,14 +269,14 @@ export class LinodeCreate extends React.PureComponent<
 
   stackScriptTabs: CreateTab[] = [
     {
-      title: 'Community StackScripts',
-      type: 'fromStackScript',
-      routeName: `${this.props.match.url}?type=StackScripts&subtype=Community`
-    },
-    {
       title: 'Account StackScripts',
       type: 'fromStackScript',
       routeName: `${this.props.match.url}?type=StackScripts&subtype=Account`
+    },
+    {
+      title: 'Community StackScripts',
+      type: 'fromStackScript',
+      routeName: `${this.props.match.url}?type=StackScripts&subtype=Community`
     }
   ];
 
@@ -452,10 +456,10 @@ export class LinodeCreate extends React.PureComponent<
                     <TabPanels className={classes.imageSelect}>
                       <SafeTabPanel index={0}>
                         <FromStackScriptContent
-                          category="community"
+                          category="account"
                           accountBackupsEnabled={accountBackupsEnabled}
                           userCannotCreateLinode={userCannotCreateLinode}
-                          request={getCommunityStackscripts}
+                          request={getMineAndAccountStackScripts}
                           header={'Select a StackScript'}
                           imagesData={imagesData!}
                           regionsData={regionsData!}
@@ -465,10 +469,10 @@ export class LinodeCreate extends React.PureComponent<
                       </SafeTabPanel>
                       <SafeTabPanel index={1}>
                         <FromStackScriptContent
-                          category="account"
+                          category="community"
                           accountBackupsEnabled={accountBackupsEnabled}
                           userCannotCreateLinode={userCannotCreateLinode}
-                          request={getMineAndAccountStackScripts}
+                          request={getCommunityStackscripts}
                           header={'Select a StackScript'}
                           imagesData={imagesData!}
                           regionsData={regionsData!}
@@ -536,7 +540,7 @@ export class LinodeCreate extends React.PureComponent<
           <SelectPlanPanel
             data-qa-select-plan
             error={hasErrorFor.type}
-            types={typesData!}
+            types={filterCurrentTypes(typesData)!}
             onSelect={this.props.updateTypeID}
             selectedID={this.props.selectedTypeID}
             updateFor={[
@@ -596,14 +600,12 @@ export class LinodeCreate extends React.PureComponent<
             privateIP={this.props.privateIPEnabled}
             changeBackups={this.props.toggleBackupsEnabled}
             changePrivateIP={this.props.togglePrivateIPEnabled}
-            updateFor={[
-              this.props.privateIPEnabled,
-              this.props.backupsEnabled,
-              this.props.selectedTypeID,
-              this.props.createType
-            ]}
             disabled={userCannotCreateLinode}
             hidePrivateIP={this.props.createType === 'fromLinode'}
+            changeSelectedVLAN={this.props.setVlanID}
+            selectedVlanID={this.props.selectedVlanID}
+            selectedRegionID={this.props.selectedRegionID}
+            vlanError={hasErrorFor.interfaces}
           />
         </Grid>
         <Grid item className="mlSidebar">
@@ -614,6 +616,11 @@ export class LinodeCreate extends React.PureComponent<
             isMakingRequest={this.props.formIsSubmitting}
             disabled={this.props.formIsSubmitting || userCannotCreateLinode}
             onDeploy={this.createLinode}
+            footer={
+              <SMTPRestrictionText>
+                {({ text }) => <div style={{ marginTop: 16 }}>{text}</div>}
+              </SMTPRestrictionText>
+            }
           >
             <DisplaySectionList displaySections={displaySections} />
           </CheckoutBar>