+++ packages/manager/src/features/linodes/LinodesDetail/LinodeResize/LinodeResize.tsx
@@ -27,12 +27,11 @@ import Typography from 'src/components/core/Typography';
 import { DocumentTitleSegment } from 'src/components/DocumentTitle';
 import ExternalLink from 'src/components/ExternalLink';
 import HelpIcon from 'src/components/HelpIcon';
+import withTypes, { WithTypesProps } from 'src/containers/types.container';
 import { resetEventsPolling } from 'src/eventsPolling';
-import SelectPlanPanel, {
-  ExtendedType
-} from 'src/features/linodes/LinodesCreate/SelectPlanPanel';
+import SelectPlanPanel from 'src/features/linodes/LinodesCreate/SelectPlanPanel';
+import { ExtendedType } from 'src/store/linodeType/linodeType.reducer';
 import { withLinodeDetailContext } from 'src/features/linodes/LinodesDetail/linodeDetailContext';
-import { typeLabelDetails } from 'src/features/linodes/presentation';
 import { linodeInTransition } from 'src/features/linodes/transitions';
 import { ApplicationState } from 'src/store';
 import { requestLinodeForStore } from 'src/store/linodes/linode.requests';
@@ -133,32 +132,11 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
     }
   };
 
-  static extendType = (type: LinodeType): ExtendedType => {
-    const {
-      label,
-      memory,
-      vcpus,
-      disk,
-      price: { monthly, hourly }
-    } = type;
-
-    const isGPU = type.class === 'gpu';
-
-    return {
-      ...type,
-      heading: label,
-      subHeadings: [
-        `$${monthly}/mo ($${isGPU ? hourly.toFixed(2) : hourly}/hr)`,
-        typeLabelDetails(memory, disk, vcpus)
-      ]
-    };
-  };
-
   openConfirmationModal = () => {
-    const { linodeType, currentTypesData } = this.props;
+    const { linodeType, typesData } = this.props;
     const { selectedId } = this.state;
-    const _current = getLinodeType(currentTypesData, linodeType || 'none');
-    const _target = getLinodeType(currentTypesData, selectedId);
+    const _current = getLinodeType(typesData, linodeType || 'none');
+    const _target = getLinodeType(typesData, selectedId);
     const currentPlan = _current ? _current.label : 'Unknown plan';
     const targetPlan = _target ? _target.label : 'Unknown plan';
     this.setState({
@@ -188,7 +166,7 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
       enqueueSnackbar,
       history,
       updateLinode,
-      currentTypesData
+      typesData
     } = this.props;
     const { selectedId } = this.state;
 
@@ -199,7 +177,7 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
     const isSmaller = isSmallerThanCurrentPlan(
       selectedId,
       linodeType || '',
-      currentTypesData
+      typesData
     );
 
     this.setState({
@@ -218,7 +196,8 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
           selectedId: '',
           confirmationDialog: {
             ...this.state.confirmationDialog,
-            submitting: false
+            submitting: false,
+            isOpen: false
           }
         });
         resetEventsPolling();
@@ -259,8 +238,7 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
 
   render() {
     const {
-      currentTypesData,
-      deprecatedTypesData,
+      typesData,
       linodeType,
       linodeLabel,
       permissions,
@@ -268,9 +246,7 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
       linodeDisks,
       linodeStatus
     } = this.props;
-    const type = [...currentTypesData, ...deprecatedTypesData].find(
-      t => t.id === linodeType
-    );
+    const type = typesData.find(t => t.id === linodeType);
 
     const hostMaintenance = linodeStatus === 'stopped';
     const unauthorized = permissions === 'read_only';
@@ -291,7 +267,7 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
     const isSmaller = isSmallerThanCurrentPlan(
       this.state.selectedId,
       linodeType || '',
-      currentTypesData
+      typesData
     );
 
     return (
@@ -323,7 +299,9 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
 
           <SelectPlanPanel
             currentPlanHeading={currentPlanHeading}
-            types={this.props.currentTypesData}
+            types={typesData.filter(
+              thisType => !thisType.isDeprecated && !thisType.isShadowPlan
+            )}
             onSelect={this.handleSelectPlan}
             selectedID={this.state.selectedId}
             disabled={disabled}
@@ -395,21 +373,6 @@ export class LinodeResize extends React.Component<CombinedProps, State> {
 }
 
 const styled = withStyles(styles);
-
-interface WithTypesProps {
-  currentTypesData: ExtendedType[];
-  deprecatedTypesData: ExtendedType[];
-}
-
-const withTypes = connect((state: ApplicationState) => ({
-  currentTypesData: state.__resources.types.entities
-    .filter(eachType => eachType.successor === null)
-    .map(LinodeResize.extendType),
-
-  deprecatedTypesData: state.__resources.types.entities
-    .filter(eachType => eachType.successor !== null)
-    .map(LinodeResize.extendType)
-}));
 
 interface DispatchProps {
   updateLinode: (id: number) => void;