+++ packages/manager/src/mocks/serverHandlers.ts
@@ -23,6 +23,7 @@ import {
   monitorFactory,
   nodeBalancerFactory,
   notificationFactory,
+  objectStorageBucketFactory,
   profileFactory,
   supportReplyFactory,
   supportTicketFactory,
@@ -72,7 +73,9 @@ export const handlers = [
     return res(ctx.json(makeResourcePage(images)));
   }),
   rest.get('*/instances', async (req, res, ctx) => {
-    const onlineLinodes = linodeFactory.buildList(3);
+    const onlineLinodes = linodeFactory.buildList(3, {
+      backups: { enabled: false }
+    });
     const offlineLinodes = linodeFactory.buildList(1, { status: 'offline' });
     const busyLinodes = linodeFactory.buildList(5, { status: 'migrating' });
     const eventLinode = linodeFactory.build({
@@ -84,6 +87,11 @@ export const handlers = [
       ...onlineLinodes,
       ...offlineLinodes,
       ...busyLinodes,
+      linodeFactory.build({
+        label: 'shadow-plan',
+        type: 'g6-standard-3-s',
+        backups: { enabled: false }
+      }),
       eventLinode
     ];
     return res(ctx.json(makeResourcePage(linodes)));
@@ -185,6 +193,10 @@ export const handlers = [
       return res(ctx.json(makeResourcePage(configs)));
     }
   ),
+  rest.get('*/object-storage/buckets/*', (req, res, ctx) => {
+    const buckets = objectStorageBucketFactory.buildList(20);
+    return res(ctx.json(makeResourcePage(buckets)));
+  }),
   rest.get('*/domains', (req, res, ctx) => {
     const domains = domainFactory.buildList(25);
     return res(ctx.json(makeResourcePage(domains)));
@@ -215,7 +227,10 @@ export const handlers = [
     return res(ctx.json(makeResourcePage(items, { page: 1, pages: 4 })));
   }),
   rest.get('*/account', (req, res, ctx) => {
-    const account = accountFactory.build({ balance: 50 });
+    const account = accountFactory.build({
+      balance: 50,
+      active_since: '2019-11-05'
+    });
     return res(ctx.json(account));
   }),
   rest.put('*/account', (req, res, ctx) => {
@@ -306,6 +321,9 @@ export const handlers = [
         ])
       )
     );
+  }),
+  rest.post('*/networking/vlans', (req, res, ctx) => {
+    return res(ctx.json({}));
   })
 ];